pyproject.toml
[tool.poetry]
name = "airq"
version = "0.1.0"
description = ""
authors = ["Jeremias Robles <jeremias.robles@nordcloud.com>"]

[tool.poetry.dependencies]
python = "^3.9.0"
pylint = "^2.6.0"
pycryptodome = "^3.9.9"
requests = "^2.24.0"

[tool.poetry.dev-dependencies]
pytest = "^4.6"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
